chore/ci: introduce Dockerfile for build
Provides a simple mechanism for building Crust inside a container. It
would primarily be used in a Travis or Jenkins-based build environment,
but it could also be used locally for developers to get an environment
that matches the build environment.

During the `docker build` process we run a `cargo build` so that all the
dependencies are downloaded and compiled for subsequent `docker run`
usage. This is why a target directory is explicitly used with Cargo.
During subsequent `docker run` usage, we will mount the current source
code in as a volume to /usr/source/cargo, which would overwrite the
default target directory used in the `docker build`.